backend/mlModel/requirements.txt
pandas
scikit-learn
psycopg2-binary
python-dotenv
sendgrid
numpy
SQLAlchemy
